Alter3 - index counter changed (now based on dictionary instead of array)

+++ Clope/ClopeLib/Data/Transaction.cs
@@ -5,11 +5,11 @@ namespace ClopeLib.Data
 {
 	public class Transaction : ITransaction
 	{
-		static uint uniqueId = 0;
+		static long uniqueId = 0;
 
 
 
-		public uint Id { get; } = uniqueId++;
+		public long Id { get; } = uniqueId++;
 
 		public int[] Links { get; }
